# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Blind Trade Engine","version":"0.1.0"},"paths":{"/api/v1/market/status":{"get":{"tags":["market"],"summary":"Get Market Status","description":"Get current market status (Open/Closed) and major index levels.","operationId":"get_market_status_api_v1_market_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/market/live/{symbol}":{"get":{"tags":["market"],"summary":"Get Live Price","description":"Get live price for a specific stock.","operationId":"get_live_price_api_v1_market_live__symbol__get","parameters":[{"name":"symbol","in":"path","required":true,"schema":{"type":"string","title":"Symbol"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/signals/today":{"get":{"tags":["signals"],"summary":"Get Todays Signals","description":"Generates signals based on Real-Time Technical Analysis.\nMode: 'intraday' (15m), 'swing' (Daily), 'longterm' (Weekly)","operationId":"get_todays_signals_api_v1_signals_today_get","parameters":[{"name":"mode","in":"query","required":false,"schema":{"type":"string","default":"swing","title":"Mode"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/signals/analyze/{symbol}":{"get":{"tags":["signals"],"summary":"Analyze Single Stock","description":"Analyze a specific stock symbol on demand.","operationId":"analyze_single_stock_api_v1_signals_analyze__symbol__get","parameters":[{"name":"symbol","in":"path","required":true,"schema":{"type":"string","title":"Symbol"}},{"name":"mode","in":"query","required":false,"schema":{"type":"string","default":"swing","title":"Mode"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/signals/sectors":{"get":{"tags":["signals"],"summary":"Get Sector Signals","description":"Returns Cached Sector Signals (Instant Response).\nStrictly separates 'intraday' vs 'longterm' from Global Cache.","operationId":"get_sector_signals_api_v1_signals_sectors_get","parameters":[{"name":"mode","in":"query","required":false,"schema":{"type":"string","default":"intraday","title":"Mode"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}